// src/api/service/model/authResponse.ts
/**
 * Things Keeper
 * Will save your notes.
 *
 * OpenAPI spec version: 0.2.0
 * Contact: vrahlinskiy@cinimex.ru
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AuthResponse { 
    token?: string;
    username?: string;
}